requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
requests
python-multipart